# [STAL-1418] Violation fingerprints