## What problem are you trying to solve?

We want to add a unique fingerprint to each violation in the sarif file. 

## What is your solution?

This is reported in the `partialFingerprint` in the SARIF file. This hash contains the content of the file and some other meta information to attempt to make it unique.


## Testing and Performance

Ran the tool on a large repo and disable the generation of fingerprints. Run on 1630 files for Go and YAML. The objective is to see the performance impact.

To disable fingerprinting, we just added `let fingerprint_option = None;` in the `sarif_utils.rs`

Then, built each version with `cargo build -r`

### Without fingerprints

```
Found 1500 violation(s) in 1630 file(s) using 105 rule(s) within 12 sec(s)
./target/release/datadog-static-analyzer --directory  --format sarif --output  57.55s user 5.52s system 435% cpu 14.486 total
```

### With fingerprints

```
Found 1500 violation(s) in 1630 file(s) using 105 rule(s) within 12 sec(s)
./target/release/datadog-static-analyzer --directory  --format sarif --output  56.25s user 5.35s system 447% cpu 13.756 total
```